requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
gnureadline==6.3.3
ipython==3.1.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
oauthlib==0.7.2
PyYAML==3.11
rauth==0.7.1
requests==2.6.0
requests-oauthlib==0.4.2
SQLAlchemy==1.0.0
Werkzeug==0.10.4
WTForms==2.0.2
